+++ domain/src/main/java/com/yesferal/hornsapp/domain/usecase/GetConcertsUseCase.kt
@@ -1,13 +1,12 @@
 package com.yesferal.hornsapp.domain.usecase
 
+import com.yesferal.hornsapp.domain.abstraction.ConcertRepository
 import com.yesferal.hornsapp.domain.entity.Concert
-import com.yesferal.hornsapp.domain.repository.ConcertRepository
 
 class GetConcertsUseCase(
     private val concertRepository: ConcertRepository
-) : BaseUseCase<Nothing, List<Concert>> {
-    override fun invoke(
-        param: Nothing?,
+) {
+    fun invoke(
         onSuccess: (response: List<Concert>) -> Unit,
         onError: (t: Throwable) -> Unit
     ) {
